# node/tsoa_build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Team": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"teamKey": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"player1Id": {
						"type": "number",
						"format": "double"
					},
					"player2Id": {
						"type": "number",
						"format": "double"
					},
					"wonGamesTotal": {
						"type": "number",
						"format": "double"
					},
					"drawGamesTotal": {
						"type": "number",
						"format": "double"
					},
					"playedGamesTotal": {
						"type": "number",
						"format": "double"
					},
					"goalsAgainst": {
						"type": "number",
						"format": "double"
					},
					"goalsFor": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"teamKey",
					"name",
					"player1Id",
					"player2Id",
					"wonGamesTotal",
					"drawGamesTotal",
					"playedGamesTotal",
					"goalsAgainst",
					"goalsFor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MatchResult": {
				"enum": [
					"win",
					"draw",
					"loss"
				],
				"type": "string"
			},
			"MatchPlayed": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"$ref": "#/components/schemas/MatchResult"
					},
					"opponentId": {
						"type": "number",
						"format": "double"
					},
					"opponentName": {
						"type": "string"
					},
					"yourScore": {
						"type": "number",
						"format": "double"
					},
					"theirScore": {
						"type": "number",
						"format": "double"
					},
					"difference": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"result",
					"opponentId",
					"opponentName",
					"yourScore",
					"theirScore",
					"difference"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Table Football Tracker API",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/teams": {
			"get": {
				"operationId": "GetTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Team"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostTeam",
				"responses": {
					"200": {
						"description": "Successfully created team!",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"description": "Could not add team. A team with the same people already exists!"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"player2Id": {},
									"player1Id": {},
									"name": {}
								},
								"required": [
									"player2Id",
									"player1Id",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/teams/{id}": {
			"get": {
				"operationId": "GetTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Team"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/teams/{id}/history": {
			"get": {
				"operationId": "GetTeamHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MatchPlayed"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}